move face search to smart info repository
