+++ server/src/domain/repositories/person.repository.ts
@@ -25,15 +25,6 @@ export interface PersonStatistics {
   assets: number;
 }
 
-export type Embedding = number[]; 
-
-export interface EmbeddingSearch {
-  ownerId: string;
-  embedding: Embedding;
-  numResults: number;
-  maxDistance?: number;
-}
-
 export interface IPersonRepository {
   getAll(): Promise<PersonEntity[]>;
   getAllWithoutThumbnail(): Promise<PersonEntity[]>;
@@ -51,7 +42,6 @@ export interface IPersonRepository {
   delete(entity: PersonEntity): Promise<PersonEntity | null>;
   deleteAll(): Promise<number>;
   getStatistics(personId: string): Promise<PersonStatistics>;
-  searchByEmbedding(search: EmbeddingSearch): Promise<AssetFaceEntity[]>;
   getAllFaces(): Promise<AssetFaceEntity[]>;
   getFacesByIds(ids: AssetFaceId[]): Promise<AssetFaceEntity[]>;
   getRandomFace(personId: string): Promise<AssetFaceEntity | null>;
