+++ server/test/repositories/smart-info.repository.mock.ts
@@ -2,7 +2,8 @@ import { ISmartInfoRepository } from '@app/domain';
 
 export const newSmartInfoRepositoryMock = (): jest.Mocked<ISmartInfoRepository> => {
   return {
-    searchByEmbedding: jest.fn(),
+    searchCLIP: jest.fn(),
+    searchFaces: jest.fn(),
     upsert: jest.fn(),
   };
 };
